Big tidyup
- Remove unnecessary configurability of check-references
- Remove category support (generally unused)
- Remove next-file (totally unused)
- Remove unnecessary configurability of minimumAcceptableAcs
