+++ README.md
@@ -106,26 +106,6 @@ npx github:vegaprotocol/approbation@latest check-references --specs="./specs/pro
 docker run -v "$(pwd):/run" ghcr.io/vegaprotocol/approbation:latest check-references --specs="/run/specs/protocol/**/*.{md,ipynb}" --tests="/run/MultisigControl/test/*.js" --ignore="/run/specs/protocol/{0001-*}" --categories="/run/specs/protocol/categories.json" --show-branches --show-mystery --output-csv --output="/run/results/"
 ```
 
-## next-filename
-> Suggests what file sequence number to use next, given a list of spec files
-
-**Arguments**
-| **Parameter**   | **Type** | **Description**                      | **Example**          |
-|-----------------|----------|--------------------------------------|----------------------|
-| `--specs`         | glob     | specs to pull AC codes from          | `{specs/**/*.md}`    |
-| `--ignore`        | glob     | glob of files not to include | `specs/0001-spec.md` |
-| `--verbose` | boolean  | Display extra output | -  |
-
-In the case of three specs, ['001...', '002...', '003...'] this would suggest '004'. However if '002' didn't exist, it would indicate that '002' is available, as well as '004'.
-### next-filename example
-```bash
-# Use node
-npx @vegaprotocol/approbation next-filename --specs="./specs/protocol/**/*.{md,ipynb}"
-
-# Or run the docker image
-docker run -v "$(pwd):/run" ghcr.io/vegaprotocol/approbation:latest next-filename --specs="/run/specs/protocol/**/*.{md,ipynb}"
-```
-
 ## next-code
 > Suggests what AC code to use next given a spec file
 
@@ -136,15 +116,15 @@ docker run -v "$(pwd):/run" ghcr.io/vegaprotocol/approbation:latest next-filenam
 | `--ignore`        | glob     | glob of files not to include | `specs/0001-spec.md` |
 | `--verbose` | boolean  | Display extra output | -  |
 
-Like `next-filename`, `next-code` will suggest the lowest available code in the sequence (e.g. if there is `0001-SPEC-001` and `0001-SPEC-003`), it will suggest both `0001-SPEC-002` and `000-SPEC-004`. If using the lowest available code, ensure it isn't already referenced by any tests (i.e. isn't listed as a 'Mystery Criteria' by `check-references`).
+`next-code` will suggest the lowest available code in the sequence (e.g. if there is `0001-SPEC-001` and `0001-SPEC-003`), it will suggest both `0001-SPEC-002` and `000-SPEC-004`. If using the lowest available code, ensure it isn't already referenced by any tests (i.e. isn't listed as a 'Mystery Criteria' by `check-references`).
 
 ### next-code example
 ```bash
 # Use node
-npx @vegaprotocol/approbation next-filename --specs="./specs/protocol/**/*.{md,ipynb}"
+npx @vegaprotocol/approbation next-code --specs="./0030-ETHM-multisig_control_spec.md"
 
 # Or run the docker image
-docker run -v "$(pwd):/run" ghcr.io/vegaprotocol/approbation:latest next-filename --specs="/run/specs/protocol/**/*.{md,ipynb}"
+docker run -v "$(pwd):/run" ghcr.io/vegaprotocol/approbation:latest next-code --specs="./0030-ETHM-multisig_control_spec.md"
 ```
