+++ src/check-references.js
@@ -10,19 +10,13 @@ const glob = require('fast-glob')
 const path = require('path')
 const pc = require('picocolors')
 const { validSpecificationPrefix, validAcceptanceCriteriaCode, ignoreFiles } = require('./lib')
-const { getCategoriesForSpec, increaseCodesForCategory, increaseCoveredForCategory, increaseAcceptableSpecsForCategory, increaseUncoveredForCategory, increaseFeatureCoveredForCategory, increaseSystemTestCoveredForCategory, increaseSpecCountForCategory, setCategories } = require('./lib/category')
-const { setFeatures, increaseAcceptableSpecsForFeature, increaseCodesForFeature, increaseCoveredForFeature, increaseFeatureCoveredForFeature, increaseSpecCountForFeature, increaseSystemTestCoveredForFeature, increaseUncoveredForFeature, specFeatures, findDuplicateAcs } = require('./lib/feature')
+const { setFeatures, increaseCoveredForFeature, increaseFeatureCoveredForFeature, increaseSystemTestCoveredForFeature, increaseUncoveredForFeature, findDuplicateAcs } = require('./lib/feature')
 const { Table } = require('console-table-printer')
-const { specPriorities } = require('./lib/priority')
 const sortBy = require('lodash.sortby')
 
 // can be overriden by passing in --current-milestone, this is set for an easy transition while Jenkins doesn't
 // have that param - see vegaprotocol/jenkins-shared-library/issues/751 
 const DEFAULT_CURRENT_MILESTONE = 'colosseo_II'
-
-// Ugly globals
-let verbose = false
-let showFiles = false
 
 function gatherSpecs(fileList) {
   // Step 1: Gather all the initial details
@@ -48,20 +42,12 @@ function gatherSpecs(fileList) {
       criteria = [...new Set(labelledAcceptanceCriteria)]
     }
 
-    let categories
-    try {
-      categories = getCategoriesForSpec(codeStart[0])
-    } catch (e) {
-      categories = [];
-    }
-
     specFiles.set(fileName, {
       name: `${path}${file}`,
       code: codeStart[0],
       file,
       path,
-      criteria,
-      categories
+      criteria
     })
   })
 
@@ -112,16 +98,9 @@ function processReferences(specs, tests) {
     value.referencedBySystemTest = 0
 
     const unreferencedCriteria = []
-    // Used as a title later on
-    const category = value.categories && value.categories !== ['Unknown'] ? ' ' + value.categories.filter(c => c !== 'Total').map(c => { return pc.bgYellow(pc.black(`#${c}`)) }).join(', ') : ''
     // Used as a shortcut for all the loops
-    const categories = value.categories
-    categories.forEach(c => increaseSpecCountForCategory(c))
-
-    const acceptableMinimum = 1
 
     criteriaTotal += value.criteria.length
-    categories.forEach(c => increaseCodesForCategory(c, value.criteria.length))
 
     let refOutput = ''
     const criteriaWithRefs = []
@@ -136,7 +115,6 @@ function processReferences(specs, tests) {
           refOutput += `${pc.green(c)}:  ${linksForAC.length} (${linksForAC.toString()})\r\n`
           criteriaWithRefs.push(c)
           criteriaReferencedTotal++
-          categories.forEach(c => increaseCoveredForCategory(c, 1))
           increaseCoveredForFeature(c, 1)
 
           // Hacky hack: Limit these to 1 or 0 rather than a true tally.
@@ -145,18 +123,14 @@ function processReferences(specs, tests) {
           linksForAC.forEach(l => {
             if (!criteriaAlreadyLoggedSystest && l.match('system-tests')) {
               increaseSystemTestCoveredForFeature(c, 1)
-              categories.forEach(c => increaseSystemTestCoveredForCategory(c, 1))
               value.referencedBySystemTest++
               criteriaAlreadyLoggedSystest = true
             } else if (!criteriaAlreadyLoggedFeature && l.match('.feature')) {
               increaseFeatureCoveredForFeature(c, 1)
-              categories.forEach(c => increaseFeatureCoveredForCategory(c, 1))
               value.referencedByFeature++
               criteriaAlreadyLoggedFeature = true
             }
           })
-        } else {
-          categories.forEach(c => increaseCoveredForCategory(c, 0))
         }
 
         // Delete used references, so that tests at the end contains only AC codes not found in specs
@@ -171,28 +145,19 @@ function processReferences(specs, tests) {
             criteriaUnreferencedTotal++
           }
         })
-
-        categories.forEach(c => increaseUncoveredForCategory(c, unreferencedCriteria.length))
-      }
-
-      if (value.criteria.length > acceptableMinimum) {
-        categories.forEach(c => increaseAcceptableSpecsForCategory(c))
       }
     }
 
-    const count = value.criteria.length > acceptableMinimum ? pc.green(value.criteria.length) : pc.red(value.criteria.length)
-    const referenced = criteriaWithRefs.length > acceptableMinimum ? pc.green(criteriaWithRefs.length) : pc.red(criteriaWithRefs.length)
+    const count = pc.green(value.criteria.length)
+    const referenced = pc.green(criteriaWithRefs.length)
 
     value.count = value.criteria.length
     value.referenced = criteriaWithRefs.length
     value.uncovered = value.count - value.referenced
-    value.priority = specPriorities[value.code] || 10
 
-    if (showFiles) {
       const tally = ` has ${count} ACs of which ${referenced} are tested`
 
-      console.log(`${pc.bold(key)}${tally}${category}`)
-      if (verbose) {
+      console.log(`${pc.bold(key)}`)
         console.group()
         console.log(`File:          ${value.file}`)
         console.log(`Unreferenced ACs: ${unreferencedCriteria.join(', ')}`)
@@ -202,8 +167,6 @@ function processReferences(specs, tests) {
           console.groupEnd()
         }
         console.groupEnd()
-      }
-    }
   })
 
   return {
@@ -220,39 +183,25 @@ function processReferences(specs, tests) {
  * 
  * @param {*} specsGlob 
  * @param {*} testsGlob 
- * @param {*} categoriesPath 
  * @param {*} ignoreGlob 
  * @param {*} featuresPath 
- * @param {*} showMystery 
- * @param {*} isVerbose 
- * @param {*} showCategoryStats 
- * @param {*} shouldShowFiles 
- * @param {*} shouldOutputCSV 
- * @param {*} shouldOutputJenkins 
- * @param {*} shouldShowFileStats 
  * @param {*} currentMilestone
  * @param {*} outputPath 
  * @returns 
  */
-function checkReferences(specsGlob, testsGlob, categoriesPath, ignoreGlob, featuresPath, showMystery = false, isVerbose = false, showCategoryStats = false, shouldShowFiles = false, shouldOutputCSV = false, shouldOutputJenkins = false, shouldShowFileStats = false, currentMilestone = DEFAULT_CURRENT_MILESTONE, outputPath = './results') {
-  verbose = isVerbose
-  showFiles = shouldShowFiles
+function checkReferences(specsGlob, testsGlob, ignoreGlob, featuresPath, currentMilestone = DEFAULT_CURRENT_MILESTONE, outputPath = './results') {
   const ignoreList = ignoreGlob ? glob.sync(ignoreGlob, {}) : []
   const specList = ignoreFiles(glob.sync(specsGlob, {}), ignoreList)
   const testList = ignoreFiles(glob.sync(testsGlob, {}), ignoreList, 'test')
-  let categories
   const milestones = new Map()
   const totals = []
 
-  let specs, tests, features, specFeatures
+  let specs, tests, specFeatures
   const exitCode = 0
 
 
   if (specList.length > 0 && testList.length > 0) {
     try {
-      // Categories gather spec files in to categories, and tally the number of codes in each category
-      specCategories = JSON.parse(fs.readFileSync(categoriesPath))
-      setCategories(specCategories)
       // Features gather Acceptance Criteria across spec files or categories, and tally the numbers
 
       if (featuresPath !== undefined && featuresPath.length > 0) {
@@ -276,33 +225,15 @@ function checkReferences(specsGlob, testsGlob, categoriesPath, ignoreGlob, featu
     const criteriaReferencedPercent = (criteriaReferencedTotal / criteriaTotal * 100).toFixed(1)
     const criteriaUnreferencedPercent = (criteriaUnreferencedTotal / criteriaTotal * 100).toFixed(1)
 
-    if (showMystery && unknownCriteriaInTests.size > 0) {
-      const g = pc.bold(`${pc.red('Mystery criteria')} referenced in tests, not found in specs:`)
-      console.group(g)
-      console.dir(unknownCriteriaInTests)
-      console.groupEnd(g)
-      console.log()
-    }
+    console.log(pc.bold('Total criteria') + `:       ${criteriaTotal}`)
+    console.log(pc.green(pc.bold('With references')) + `:      ${criteriaReferencedTotal} (${criteriaReferencedPercent}%)`)
+    console.log(pc.red(pc.bold('Without references')) + `:   ${criteriaUnreferencedTotal} (${criteriaUnreferencedPercent}%)`)
 
-    if (!showCategoryStats) {
-      console.log(pc.bold('Total criteria') + `:       ${criteriaTotal}`)
-      console.log(pc.green(pc.bold('With references')) + `:      ${criteriaReferencedTotal} (${criteriaReferencedPercent}%)`)
-      console.log(pc.red(pc.bold('Without references')) + `:   ${criteriaUnreferencedTotal} (${criteriaUnreferencedPercent}%)`)
-    }
-    if (showMystery) {
-      console.log(pc.red(pc.bold('Mystery criteria')) + `:     ${unknownCriteriaInTests.size}`)
-    }
-
-    let specsTableRows
-
-    if (shouldShowFileStats) {
-      specsTableRows = Array.from(specs.keys()).map(key => {
+    let specsTableRows = Array.from(specs.keys()).map(key => {
         const s = specs.get(key)
         const coverage = (s.referenced / s.count * 100).toFixed(1)
         return {
           File: key,
-          Priority: s.priority,
-          Category: s.category,
           Criteria: s.count,
           Covered: s.referenced,
           'by/FeatTest': s.referencedByFeature,
@@ -312,9 +243,8 @@ function checkReferences(specsGlob, testsGlob, categoriesPath, ignoreGlob, featu
         }
       })
       const st = new Table()
-      st.addRows(sortBy(specsTableRows, ['Priority', 'Coverage']))
+      st.addRows(sortBy(specsTableRows, ['Coverage']))
       console.log(st.render())
-    }
 
     if (featuresPath) {
       const milestoneNames = new Set()
@@ -326,24 +256,13 @@ function checkReferences(specsGlob, testsGlob, categoriesPath, ignoreGlob, featu
         const c = specFeatures[key]
         const coverage = (c.covered / (c.acs.length | 0) * 100).toFixed(1)
 
-        const mysteryFeatureAcs = []
-        c.acs.forEach(ac => {
-          if (allCriteriaInSpecs.indexOf(ac) === -1) {
-            mysteryFeatureAcs.push(ac)
-          }
-        })
-
         const duplicateAcs = findDuplicateAcs(c.acs);
-        if (duplicateAcs.length > 0 || mysteryFeatureAcs.length > 0 || c.uncovered !== 0) {
+        if (duplicateAcs.length > 0 || c.uncovered !== 0) {
           console.group(pc.bold(`Feature errors: ${key}`))
 
           console.group();
           if (duplicateAcs.length > 0) {
             console.log(pc.red(pc.bold(`Duplicate ACs for ${key}(${c.milestone})`)) + ` ${duplicateAcs.join(', ')}`)
-          }
-
-          if (mysteryFeatureAcs.length > 0) {
-            console.log(pc.red(pc.bold(`Mystery ACs for ${key}(${c.milestone})`)) + ` ${JSON.stringify(mysteryFeatureAcs)} `)
           }
 
           if (c.uncovered !== 0 && c.uncoveredAcs) {
@@ -401,53 +320,10 @@ function checkReferences(specsGlob, testsGlob, categoriesPath, ignoreGlob, featu
       console.log(tableOutput)
     }
 
-    if (showCategoryStats) {
-      const shouldOutputImage = false
-
-      categories = Object.keys(specCategories).map(key => {
-        const c = specCategories[key]
-        const coverage = (c.covered / c.codes * 100).toFixed(1)
-
-        return {
-          Category: key,
-          Specs: c.specCount || '-',
-          Acceptable: c.acceptableSpecCount ? c.acceptableSpecCount : '-',
-          Criteria: c.codes || '-',
-          Covered: c.covered || '-',
-          'by/FeatTest': c.featureCovered || '-',
-          'by/SysTest': c.systemTestCovered || '-',
-          Uncovered: c.uncovered || '-',
-          Coverage: isNaN(coverage) ? '-' : `${coverage}% `
-        }
-      })
-
-      const t = new Table()
-      t.addRows(categories)
-      const tableOutput = t.render()
-      console.log(tableOutput)
-
-      if (shouldOutputJenkins || shouldOutputCSV || shouldOutputImage) {
         if (!fs.existsSync(outputPath)) {
           fs.mkdirSync(outputPath, { recursive: true })
         }
 
-        if (shouldOutputCSV) {
-          let categoriesCsvOutput = Object.keys(categories[0]).join(',')
-          categories.forEach(c => {
-            categoriesCsvOutput += `\r\n${Object.values(c).join(',')} `
-          })
-          fs.writeFileSync(`${outputPath}/approbation-categories.csv`, categoriesCsvOutput)
-
-          if (shouldShowFileStats) {
-            let csvOutputFiles = Object.keys(specsTableRows[0]).join(',')
-            specsTableRows.forEach(c => {
-              csvOutputFiles += `\r\n${Object.values(c).join(',')}`
-            })
-            fs.writeFileSync(`${outputPath}/approbation-files.csv`, csvOutputFiles)
-          }
-        }
-
-        if (shouldOutputJenkins) {
           let cm
           // If current milestone is set by parameter, use it
           if (currentMilestone) {
@@ -458,19 +334,14 @@ function checkReferences(specsGlob, testsGlob, categoriesPath, ignoreGlob, featu
             cm = totals && totals.length > 0 ? totals.pop() : false
           }
           
-          const skipCategories = ['Category', 'Specs', 'Acceptable']
-          let jenkinsLine = `All ACs: ${Object.entries(categories.pop()).map(([key, value]) => skipCategories.indexOf(key) === -1 ? `*${key}*: ${value}` : '').join('  ').trim()}`
-
+          let jenkinsLine = `Total ACs: ${criteriaTotal}, Referenced ACs: ${criteriaReferencedTotal}, Unreferenced ACs: ${criteriaUnreferencedTotal}, Coverage: ${criteriaReferencedPercent}%`
           if (cm && cm.Milestone && cm.Coverage) {
             jenkinsLine += `\r\nCurrent milestone ACs: *${cm.Milestone}*: ${cm.Coverage}`
           }
           
           fs.writeFileSync(`${outputPath}/jenkins.txt`, jenkinsLine)
-        }
 
         console.groupEnd()
-      }
-    }
 
     return {
       exitCode,
@@ -481,7 +352,6 @@ function checkReferences(specsGlob, testsGlob, categoriesPath, ignoreGlob, featu
         criteriaReferencedPercent,
         criteriaUnreferencedPercent,
         unknownCriteriaInTests,
-        categories
       }
     }
   } else {
