--- src/config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  // What is the acceptable minimum Acceptance Criteria count for a specification?
-  minimumAcceptableACsPerSpec: 1
-}
