--- src/lib/priority.js
@@ -1,73 +1,0 @@
-// This should be committed as a JSON file to the specs repo - but for now, this will do
-const specPriorities = {
-  '0004-AMND': 1,
-  '0022-AUTH': 1,
-  '0066-VALW': 1,
-  '0069-VCBS': 1,
-  '0077-SNAP': 1,
-  '0073-LIMN': 1,
-  '0074-BTCH': 1,
-  '0075-PLUP': 1,
-  '0076-DANO': 1,
-  // ==============
-  '0006-POSI': 2,
-  '0008-TRAD': 2,
-  '0016-PFUT': 2,
-  '0027-ASSP': 2,
-  '0028-GOVE': 2,
-  '0033-OCAN': 2,
-  '0045-DSRC': 2,
-  '0046-DSRM': 2,
-  '0047-DSRF': 2,
-  '0048-DSRI': 2,
-  '0052-FPOS': 2,
-  '0055-TREA': 2,
-  '0056-REWA': 2,
-  '0057-TRAN': 2,
-  '0061-REWP': 2,
-  '0062-SPAM': 2,
-  '0064-VALP': 2,
-  '0067-KEYS': 2,
-  '0072-SPPW': 2,
-  '0078-NWLI': 2,
-  '0079-TGAP': 2,
-  '0011-NP-CLIE': 2,
-  '0063-VALK': 2,
-  '0012-NP-LIPE': 2,
-  '0080-SPOT': 2,
-  // ==============
-  '0017-PART': 3,
-  '0025-OCRE': 3,
-  '0044-LIME': 3,
-  '0070-MKTD': 3,
-  // ==============
-  '0021-MDAT': 4,
-  '0038-OLIQ': 4,
-  // ==============
-  '0002-STTL': 5,
-  '0010-MARG': 5,
-  '0011-MARA': 5,
-  '0019-MCAL': 5,
-  '0039-MKTD': 5,
-  '0040-ASSF': 5,
-  '0041-TSTK': 5,
-  '0042-LIQF': 5,
-  // ==============
-  '0026-AUCT': 6,
-  '0043-MKTL': 6,
-  // ==============
-  '0024-OSTA': 7,
-  '0051-PROD': 7,
-  // ==============
-  '0007-POSN': 8,
-  '0036-BRIE': 8,
-  '0054-NETP': 8,
-  // ==============
-  '0013-ACCT': 9,
-  '0015-INSR': 9,
-  '0018-RSKM': 9
-}
-
-module.exports = {
-  specPriorities
-}
