--- test/next-filename/missing-sequence/0001-VALI-valid_filename.md
@@ -1,1 +1,0 @@
-Lorem ipsum dolor sit amet, consectetur adipiscing elit. Praesent lobortis eros sit amet justo eleifend, eget egestas nulla consectetur. Phasellus id sapien ut nibh auctor viverra. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque nec odio volutpat, cursus augue sit amet, auctor urna. Nulla tincidunt blandit quam ut convallis. Sed mollis felis id leo lacinia, non blandit metus commodo. Curabitur ullamcorper ultrices aliquam. Nam eu velit diam. Sed id vehicula ligula. Fusce tempus nec nunc sed viverra. Vivamus a nisl nibh. Mauris tempus lobortis malesuada. Phasellus non quam nisi.
