--- test/next-filename/missing-sequence/0003-DONE-donec_euismod.md
@@ -1,1 +1,0 @@
-Donec sagittis tempor aliquam. Vestibulum ut ex et erat iaculis mollis. Maecenas congue congue odio, ac placerat dui facilisis eget. Suspendisse id augue finibus magna aliquam consequat eu sed nulla. Aliquam a porttitor felis, sed semper nulla. Quisque mattis fermentum erat, ac ultricies mi tempor sed. Aenean blandit metus vel vestibulum imperdiet.
